requirements.txt
pytest
Flask
psycopg2-binary
python-dotenv

